remove html2canvas and manually add svg overlays for exportViewToJPG
the version of html2canvas that we were using was not properly working on
firefox and safari. After investigating this further I realized that using
the latest version with foreignObjectRendering will fix the issue in firefox.

In the end I decided to manually add the svg overlays myself. it was a lot easier
and now we're in control of how this works and it's not just a black-box.