Correcting how ignnored attributes are handled
When reading a file, if it has the attribute present in _addedProperties, then the value of the attribute is also check. If the value in the file matches the one that in _addedProperties, then we dont need to add that attribute to the output. But if the attribute is present in the file but the value does not matches, then we have to add the attribute to the output to make sure that the component appears properly. The original value is added to _ignoredAttributes, and since while writing the output, the _ignoredAttributes is read second, it overrides the original value which was copied from _attrs, thus maintaining the consistency of the attributes in the output file.

For this approach to work it is necesasry that  _ignoredAttributes, are read last when writing the output file.