fix channel filters
After discussing about the filters I realized what we're doing is
completely wrong. We're not supposed to change the R G B values
for contrast/brightness/gamma, and instead we should modify the
value (v in hsv format).

I also made sure that we're doing this once so we don't have to
go over the pixels multiple times.