fix a bug that was causing some confusion in the UI
when hue or saturation change, we're changing the greyscale as well
to ensure the hue/saturation chagne is getting reflected.

But when we're changing displayGreyscale at the same as hue/saturation,
we have to make sure the displayGreyscale change is applied at last.
This will ensure that the end result is based on actual displayGreyscale
input and not state of hue/saturation