properly set the width/height of the p tag for text drawing tool
when we're parsing the SVG to show it on the osd-viewer page, we
should use fit-content for width and height as the foreignObject itself
has the proper width, height, x, and y.

but when we're creating the p tag from the textarea we should borrow
its height/width and later these values are adjusted and moved to the
foreignObject itself.

without this change the text hover wasn't only based on the visible text
and was dependent on the width of the textarea when users were drawing the text